move NotificationSettingsRepository to NotificationSettingsService
