+++ HwProj.NotificationsService/HwProj.NotificationsService.API/EventHandlers/NewCourseMateHandler.cs
@@ -14,14 +14,14 @@ namespace HwProj.NotificationsService.API.EventHandlers
     public class NewCourseMateHandler : EventHandlerBase<NewCourseMateEvent>
     {
         private readonly INotificationsRepository _notificationRepository;
-        private readonly INotificationSettingsRepository _settingsRepository;
+        private readonly INotificationSettingsService _settingsService;
         private readonly IAuthServiceClient _authServiceClient;
         private readonly IConfigurationSection _configuration;
         private readonly IEmailService _emailService;
 
         public NewCourseMateHandler(
             INotificationsRepository notificationRepository,
-            INotificationSettingsRepository settingsRepository,
+            INotificationSettingsService settingsService,
             IAuthServiceClient authServiceClient,
             IConfiguration configuration,
             IEmailService emailService)
@@ -29,7 +29,7 @@ namespace HwProj.NotificationsService.API.EventHandlers
             _notificationRepository = notificationRepository;
             _authServiceClient = authServiceClient;
             _emailService = emailService;
-            _settingsRepository = settingsRepository;
+            _settingsService = settingsService;
             _configuration = configuration.GetSection("Notification");
         }
 
@@ -44,7 +44,7 @@ namespace HwProj.NotificationsService.API.EventHandlers
             //TODO: fix
             foreach (var mentor in mentors)
             {
-                var setting = await _settingsRepository.GetAsync(mentor.UserId,
+                var setting = await _settingsService.GetAsync(mentor.UserId,
                     NotificationsSettingCategory.NewCourseMateCategory);
                 if (!setting!.IsEnabled) continue;
