+++ HwProj.NotificationsService/HwProj.NotificationsService.API/EventHandlers/StudentPassTaskEventHandler.cs
@@ -15,17 +15,17 @@ namespace HwProj.NotificationsService.API.EventHandlers
         private readonly IAuthServiceClient _authServiceClient;
         private readonly IConfigurationSection _configuration;
         private readonly IEmailService _emailService;
-        private readonly INotificationSettingsRepository _settingsRepository;
+        private readonly INotificationSettingsService _settingsService;
 
         public StudentPassTaskEventHandler(
             IAuthServiceClient authServiceClient,
             IConfiguration configuration,
             IEmailService emailService,
-            INotificationSettingsRepository settingsRepository)
+            INotificationSettingsService settingsService)
         {
             _authServiceClient = authServiceClient;
             _emailService = emailService;
-            _settingsRepository = settingsRepository;
+            _settingsService = settingsService;
             _configuration = configuration.GetSection("Notification");
         }
 
@@ -38,7 +38,7 @@ namespace HwProj.NotificationsService.API.EventHandlers
 
             foreach (var mentorId in @event.Course.MentorIds)
             {
-                var setting = await _settingsRepository.GetAsync(mentorId,
+                var setting = await _settingsService.GetAsync(mentorId,
                     NotificationsSettingCategory.NewSolutionsCategory);
                 if (!setting!.IsEnabled) continue;
