+++ HwProj.NotificationsService/HwProj.NotificationsService.API/Startup.cs
@@ -31,7 +31,7 @@ namespace HwProj.NotificationsService.API
             var connectionString = ConnectionString.GetConnectionString(Configuration);
             services.AddDbContext<NotificationsContext>(options => options.UseSqlServer(connectionString));
             services.AddScoped<INotificationsRepository, NotificationsRepository>();
-            services.AddScoped<INotificationSettingsRepository, NotificationSettingsRepository>();
+            services.AddScoped<INotificationSettingsService, NotificationSettingsService>();
             services.AddEventBus(Configuration);
             services.AddTransient<IEventHandler<StudentRegisterEvent>, RegisterEventHandler>();
             services.AddTransient<IEventHandler<RateEvent>, RateEventHandler>();
